+++ cohort-lettuce/src/main/kotlin/com/sksamuel/cohort/redis/RedisConnectionWarmup.kt
@@ -7,7 +7,7 @@ import kotlin.random.Random
 import kotlin.time.Duration
 import kotlin.time.Duration.Companion.milliseconds
 
-class RedisConnectionWarmupHealthCheck(
+class RedisConnectionWarmup(
    client: RedisClient,
    override val iterations: Int = 2500,
    override val interval: Duration = 2.milliseconds,